+++ example/app/web/flutter_service_worker.js
@@ -4,9 +4,9 @@ const TEMP = 'flutter-temp-cache';
 const CACHE_NAME = 'flutter-app-cache';
 
 const RESOURCES = {"version.json": "11e1a177d57c5da2bc6e3a6acc3010f4",
-"index.html": "a77fee5bd4bcd4a68b04ea8c2084012d",
-"/": "a77fee5bd4bcd4a68b04ea8c2084012d",
-"main.dart.js": "400327d442f5d6c2efc41e8f168607b9",
+"index.html": "027d4915b5650b57ecef3657613da765",
+"/": "027d4915b5650b57ecef3657613da765",
+"main.dart.js": "22b9d73ff5104fe37e7cf30d2efaacc4",
 "flutter.js": "6fef97aeca90b426343ba6c5c9dc5d4a",
 "favicon.png": "7b1e138b59ab213ab88128833a2681f3",
 "icons/Icon-192.png": "786a784711f7ba2dcc1aa8f4fa52d02b",